requirements.txt
scikit-learn
pandas
matplotlib
numpy
sf-hamilton